# [OpenGL] [perf] Support ti.block_dim as block size hint