--- a/tests/python/test_parallel_range_for.py
+++ b/tests/python/test_parallel_range_for.py
@@ -1,7 +1,8 @@
 import taichi as ti
 
 
-@ti.all_archs
+# such small block_dim will cause grid_dim too large for OpenGL...
+@ti.archs_excluding(ti.opengl)
 def test_parallel_range_for():
     n = 1024 * 1024
     val = ti.var(ti.i32, shape=(n))